+++ src/component/Footer/FooterUpperImage.js
@@ -1,0 +1,16 @@
+import React, { Component } from "react";
+
+class FooterUpperImage extends Component {
+  render() {
+    return (
+      <div className="iconWesh">
+        <img
+          src="http://img.lush.co.kr/images/core/footer-icons.svg"
+          alt="러쉬 하단 이미지"
+        />
+      </div>
+    );
+  }
+}
+
+export default FooterUpperImage;
